Big cargo(before)
In the first step of the project, suppose we received a large shipment of foreign chocolates and fortunately, the seller has provided us with a list of the complete specifications of each chocolate. Now we want to further examine these specifications and improve some of the storage structure of this information.