Selling chocolate
In this step of the project, we are going to set a price for chocolates according to some of their specifications and the research we have done. For this, we use the data that we have processed in the previous step and then we try to improve our price estimation step by step.