Better chocolates
Suppose that after trying some regular (non-dark) chocolates, we realize that many of them are also delicious, and if we can find better quality ones, they will probably sell well. So, in the last step of the project, we want to identify companies that produce higher quality chocolates and separate their chocolates.